Removed an example. Fixed some bugs.
Fixed a bug in the textprocessing_encode_chr (unicode encoder) function's UTF-8 encoder.
The bug was that it didn't reset the current byte when OR-ing the byte-count-indicator on a byte seperate from the last byte.

The stdin echo example can now input and echo unicode characters higher than U+10000
Previously it wasn't able to. I didn't realize that it wasn't able to so.
This was achieved by using SetConsoleOutputCP to CP_UTF8. Previously it just did _setmode to UTF_16.
Because _setmode needed you to always use wide print functions with the file, printing anything on Windows used to always use wide print functions
Now that I switched it to code pages, it can now print with normal narrow print functions.

Also removed the windows-only stdin echo example because there's the winnix one why have one with just Windows.

Also made it so that the part of textreader_get_byte that uses fgetwc store the result in a wint_t so that it can fit a WEOF and check for errors before setting it to the wchar buffer

I probably also did some other stuff but I don't remember and I'm too lazy to list them all so yeah.
The one I listed above should already contain some of the biggest (even if they're not big) ones though.